+++ assets/javascripts/kitten/components/structure/cards/manager-card/stories.js
@@ -9,8 +9,33 @@ import {
   EllipsisIcon,
   FlexWrapper,
   LockOutlineIcon,
-  AvatarWithTextAndBadge,
 } from 'kitten'
+
+const args = {
+  as: 'article',
+  actionProps: {
+    href: '#',
+    as: 'a',
+    'aria-label': 'A label for the link',
+    onClick: e => {
+      e.preventDefault()
+      action('Clicked')(e)
+    },
+  },
+}
+
+const argTypes = {
+  as: {
+    name: 'as',
+    description: 'HTML element for the Action component',
+    control: { type: 'text' },
+  },
+  actionProps: {
+    name: 'actionProps',
+    description: 'Properties for the Action component',
+    control: { type: 'object' },
+  },
+}
 
 export default {
   title: 'Structure/Cards/ManagerCard',
@@ -27,13 +52,15 @@ export default {
       </div>
     ),
   ],
+  args: args,
+  argTypes: argTypes,
 }
 
 const StyledFlexWrapper = styled(FlexWrapper)`
   height: 100%;
 `
 
-export const WithAction = args => {
+export const Default = args => {
   return (
     <ManagerCard {...args}>
       <ManagerCard.Cell>
@@ -88,90 +115,4 @@ export const WithAction = args => {
       </ManagerCard.LastCell>
     </ManagerCard>
   )
-}
-WithAction.args = {
-  as: 'article',
-  actionProps: {
-    href: '#',
-    as: 'a',
-    'aria-label': 'A label for the link',
-    onClick: e => {
-      e.preventDefault()
-      action('Clicked')(e)
-    },
-  },
-}
-WithAction.argTypes = {
-  as: {
-    name: 'as',
-    description: 'HTML element for the Action component',
-    control: { type: 'text' },
-  },
-  actionProps: {
-    name: 'actionProps',
-    description: 'Properties for the Action component',
-    control: { type: 'object' },
-  },
-}
-
-export const WithAvatar = args => {
-  return (
-    <ManagerCard {...args}>
-      <ManagerCard.Image>
-        <AvatarWithTextAndBadge>
-          <AvatarWithTextAndBadge.Image alt="" src="/kitten-0.jpg" />
-        </AvatarWithTextAndBadge>
-      </ManagerCard.Image>
-      <ManagerCard.Cell>
-        <StyledFlexWrapper gap={8} className="k-u-flex-justifyContent-center">
-          <Text
-            tag="p"
-            weight="bold"
-            size="large"
-            className="k-u-line-height-1 k-u-margin-none k-u-clamp-1 k-u-clamp-2@xs-down"
-          >
-            Marie Bobo
-          </Text>
-          <Text
-            tag="p"
-            weight="light"
-            size="micro"
-            className="k-u-line-height-1 k-u-margin-none"
-          >
-            marie.bo@gmail.com
-          </Text>
-        </StyledFlexWrapper>
-      </ManagerCard.Cell>
-      <ManagerCard.LastCell>
-        <DropdownMenu
-          menuPosition="left"
-          button={({ open }) => (
-            <>
-              <EllipsisIcon color="var(--color-grey-900)" />
-              <span className="k-u-a11y-visuallyHidden">
-                {open ? 'Click to close menu' : 'Click to open menu'}
-              </span>
-            </>
-          )}
-        >
-          <DropdownMenu.Link href="#">Editer</DropdownMenu.Link>
-          <DropdownMenu.Link href="#">Dupliquer</DropdownMenu.Link>
-          <DropdownMenu.Link href="#">Désactiver</DropdownMenu.Link>
-          <DropdownMenu.Button>Supprimer</DropdownMenu.Button>
-        </DropdownMenu>
-      </ManagerCard.LastCell>
-    </ManagerCard>
-  )
-}
-
-WithAvatar.args = {
-  as: 'article',
-}
-
-WithAvatar.argTypes = {
-  as: {
-    name: 'as',
-    description: 'HTML element for the Action component',
-    control: { type: 'text' },
-  },
 }