+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: `ManagerCard`: Add `ManagerCard.Image` subcomponent.
+- Feature: `PeopleCard`: Add new component.
 
 ## [10.1.0] - 2022-04-08